[RAM] Fix failing rules list bulk action jest tests (#142680)
## Summary
Resolves: https://github.com/elastic/kibana/issues/141052

Refactors the rules list test file to help prevent it from timing out
when running in CI

### Checklist
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit a9f801c9196e2eee609a55547da63286ff23f0bc)
