+++ x-pack/plugins/triggers_actions_ui/public/application/sections/rules_list/components/rules_list.test.tsx
@@ -12,19 +12,21 @@ import { ReactWrapper } from 'enzyme';
 import { actionTypeRegistryMock } from '../../../action_type_registry.mock';
 import { ruleTypeRegistryMock } from '../../../rule_type_registry.mock';
 import { RulesList, percentileFields } from './rules_list';
-import { RuleTypeModel, ValidationResult, Percentiles } from '../../../../types';
-import {
-  RuleExecutionStatusErrorReasons,
-  RuleExecutionStatusWarningReasons,
-  ALERTS_FEATURE_ID,
-  parseDuration,
-} from '@kbn/alerting-plugin/common';
+import { RuleTypeModel, Percentiles } from '../../../../types';
+import { parseDuration } from '@kbn/alerting-plugin/common';
 import { getFormattedDuration, getFormattedMilliseconds } from '../../../lib/monitoring_utils';
 import { getIsExperimentalFeatureEnabled } from '../../../../common/get_experimental_features';
 
 import { useKibana } from '../../../../common/lib/kibana';
 import { __IntlProvider as IntlProvider } from '@kbn/i18n-react';
 import { IToasts } from '@kbn/core/public';
+
+import {
+  mockedRulesData,
+  ruleTypeFromApi,
+  getDisabledByLicenseRuleTypeFromApi,
+  ruleType,
+} from './test_helpers';
 
 jest.mock('../../../../common/lib/kibana');
 jest.mock('@kbn/kibana-react-plugin/public/ui_settings/use_ui_setting', () => ({
@@ -80,14 +82,7 @@ jest.mock('../../../../common/get_experimental_features', () => ({
 
 const ruleTags = ['a', 'b', 'c', 'd'];
 
-const {
-  loadRuleTypes,
-  updateAPIKey,
-  loadRuleTags,
-  bulkSnoozeRules,
-  bulkUnsnoozeRules,
-  bulkUpdateAPIKey,
-} = jest.requireMock('../../../lib/rule_api');
+const { loadRuleTypes, updateAPIKey, loadRuleTags } = jest.requireMock('../../../lib/rule_api');
 const { loadRuleAggregationsWithKueryFilter } = jest.requireMock(
   '../../../lib/rule_api/aggregate_kuery_filter'
 );
@@ -96,238 +91,11 @@ const { loadActionTypes, loadAllActions } = jest.requireMock('../../../lib/actio
 
 const actionTypeRegistry = actionTypeRegistryMock.create();
 const ruleTypeRegistry = ruleTypeRegistryMock.create();
-const ruleType = {
-  id: 'test_rule_type',
-  description: 'test',
-  iconClass: 'test',
-  documentationUrl: null,
-  validate: (): ValidationResult => {
-    return { errors: {} };
-  },
-  ruleParamsExpression: () => null,
-  requiresAppContext: false,
-};
-const ruleTypeFromApi = {
-  id: 'test_rule_type',
-  name: 'some rule type',
-  actionGroups: [{ id: 'default', name: 'Default' }],
-  recoveryActionGroup: { id: 'recovered', name: 'Recovered' },
-  actionVariables: { context: [], state: [] },
-  defaultActionGroupId: 'default',
-  producer: ALERTS_FEATURE_ID,
-  minimumLicenseRequired: 'basic',
-  enabledInLicense: true,
-  authorizedConsumers: {
-    [ALERTS_FEATURE_ID]: { read: true, all: true },
-  },
-  ruleTaskTimeout: '1m',
-};
+
 ruleTypeRegistry.list.mockReturnValue([ruleType]);
 actionTypeRegistry.list.mockReturnValue([]);
 
 const useKibanaMock = useKibana as jest.Mocked<typeof useKibana>;
-
-const mockedRulesData = [
-  {
-    id: '1',
-    name: 'test rule',
-    tags: ['tag1'],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '1s' },
-    actions: [],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'active',
-      lastDuration: 500,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      error: null,
-    },
-    monitoring: {
-      execution: {
-        history: [
-          {
-            success: true,
-            duration: 1000000,
-          },
-          {
-            success: true,
-            duration: 200000,
-          },
-          {
-            success: false,
-            duration: 300000,
-          },
-        ],
-        calculated_metrics: {
-          success_ratio: 0.66,
-          p50: 200000,
-          p95: 300000,
-          p99: 300000,
-        },
-      },
-    },
-  },
-  {
-    id: '2',
-    name: 'test rule ok',
-    tags: ['tag1'],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '5d' },
-    actions: [],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'ok',
-      lastDuration: 61000,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      error: null,
-    },
-    monitoring: {
-      execution: {
-        history: [
-          {
-            success: true,
-            duration: 100000,
-          },
-          {
-            success: true,
-            duration: 500000,
-          },
-        ],
-        calculated_metrics: {
-          success_ratio: 1,
-          p50: 0,
-          p95: 100000,
-          p99: 500000,
-        },
-      },
-    },
-  },
-  {
-    id: '3',
-    name: 'test rule pending',
-    tags: ['tag1'],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '5d' },
-    actions: [],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'pending',
-      lastDuration: 30234,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      error: null,
-    },
-    monitoring: {
-      execution: {
-        history: [{ success: false, duration: 100 }],
-        calculated_metrics: {
-          success_ratio: 0,
-        },
-      },
-    },
-  },
-  {
-    id: '4',
-    name: 'test rule error',
-    tags: ['tag1'],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '5d' },
-    actions: [{ id: 'test', group: 'rule', params: { message: 'test' } }],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'error',
-      lastDuration: 122000,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      error: {
-        reason: RuleExecutionStatusErrorReasons.Unknown,
-        message: 'test',
-      },
-    },
-  },
-  {
-    id: '5',
-    name: 'test rule license error',
-    tags: [],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '5d' },
-    actions: [{ id: 'test', group: 'rule', params: { message: 'test' } }],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'error',
-      lastDuration: 500,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      error: {
-        reason: RuleExecutionStatusErrorReasons.License,
-        message: 'test',
-      },
-    },
-  },
-  {
-    id: '6',
-    name: 'test rule warning',
-    tags: [],
-    enabled: true,
-    ruleTypeId: 'test_rule_type',
-    schedule: { interval: '5d' },
-    actions: [{ id: 'test', group: 'rule', params: { message: 'test' } }],
-    params: { name: 'test rule type name' },
-    scheduledTaskId: null,
-    createdBy: null,
-    updatedBy: null,
-    apiKeyOwner: null,
-    throttle: '1m',
-    muteAll: false,
-    mutedInstanceIds: [],
-    executionStatus: {
-      status: 'warning',
-      lastDuration: 500,
-      lastExecutionDate: new Date('2020-08-20T19:23:38Z'),
-      warning: {
-        reason: RuleExecutionStatusWarningReasons.MAX_EXECUTABLE_ACTIONS,
-        message: 'test',
-      },
-    },
-  },
-];
 
 beforeEach(() => {
   (getIsExperimentalFeatureEnabled as jest.Mock<any, any>).mockImplementation(() => false);
@@ -1813,23 +1581,7 @@ describe('rules_list with disabled items', () => {
       },
     ]);
 
-    loadRuleTypes.mockResolvedValue([
-      ruleTypeFromApi,
-      {
-        id: 'test_rule_type_disabled_by_license',
-        name: 'some rule type that is not allowed',
-        actionGroups: [{ id: 'default', name: 'Default' }],
-        recoveryActionGroup: { id: 'recovered', name: 'Recovered' },
-        actionVariables: { context: [], state: [] },
-        defaultActionGroupId: 'default',
-        producer: ALERTS_FEATURE_ID,
-        minimumLicenseRequired: 'platinum',
-        enabledInLicense: false,
-        authorizedConsumers: {
-          [ALERTS_FEATURE_ID]: { read: true, all: true },
-        },
-      },
-    ]);
+    loadRuleTypes.mockResolvedValue([ruleTypeFromApi, getDisabledByLicenseRuleTypeFromApi()]);
     loadAllActions.mockResolvedValue([]);
 
     ruleTypeRegistry.has.mockReturnValue(false);
@@ -1879,233 +1631,5 @@ describe('rules_list with disabled items', () => {
     wrapper.find('[data-test-subj="rulesListNotifyBadge-unsnoozed"]').first().simulate('click');
 
     expect(wrapper.find('[data-test-subj="snoozePanel"]').exists()).toBeTruthy();
-  });
-});
-
-// Failing: https://github.com/elastic/kibana/issues/141052
-describe.skip('Rules list bulk actions', () => {
-  let wrapper: ReactWrapper<any>;
-
-  async function setup(authorized: boolean = true) {
-    loadRulesWithKueryFilter.mockResolvedValue({
-      page: 1,
-      perPage: 10000,
-      total: 6,
-      data: mockedRulesData,
-    });
-
-    loadActionTypes.mockResolvedValue([
-      {
-        id: 'test',
-        name: 'Test',
-      },
-      {
-        id: 'test2',
-        name: 'Test2',
-      },
-    ]);
-    loadRuleTypes.mockResolvedValue([
-      ruleTypeFromApi,
-      {
-        id: 'test_rule_type_disabled_by_license',
-        name: 'some rule type that is not allowed',
-        actionGroups: [{ id: 'default', name: 'Default' }],
-        recoveryActionGroup: { id: 'recovered', name: 'Recovered' },
-        actionVariables: { context: [], state: [] },
-        defaultActionGroupId: 'default',
-        producer: ALERTS_FEATURE_ID,
-        minimumLicenseRequired: 'platinum',
-        enabledInLicense: false,
-        authorizedConsumers: {
-          [ALERTS_FEATURE_ID]: { read: true, all: authorized },
-        },
-      },
-    ]);
-    loadAllActions.mockResolvedValue([]);
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    useKibanaMock().services.ruleTypeRegistry = ruleTypeRegistry;
-
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    useKibanaMock().services.actionTypeRegistry = actionTypeRegistry;
-    wrapper = mountWithIntl(<RulesList />);
-
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-  }
-
-  it('renders select all button for bulk editing', async () => {
-    await setup();
-    expect(wrapper.find('[data-test-subj="totalRulesCount"]').exists()).toBeTruthy();
-    expect(wrapper.find('[data-test-subj="showBulkActionButton"]').exists()).toBeFalsy();
-    expect(wrapper.find('[data-test-subj="selectAllRulesButton"]').exists()).toBeFalsy();
-
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-
-    expect(wrapper.find('[data-test-subj="totalRulesCount"]').exists()).toBeFalsy();
-    expect(wrapper.find('[data-test-subj="showBulkActionButton"]').exists()).toBeTruthy();
-    expect(wrapper.find('[data-test-subj="selectAllRulesButton"]').exists()).toBeTruthy();
-  });
-
-  it('does not render select all button if the user is not authorized', async () => {
-    await setup(false);
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    expect(wrapper.find('[data-test-subj="showBulkActionButton"]').exists()).toBeTruthy();
-    expect(wrapper.find('[data-test-subj="selectAllRulesButton"]').exists()).toBeFalsy();
-  });
-
-  it('selects all will select all items', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-
-    mockedRulesData.forEach((rule) => {
-      expect(
-        wrapper.find(`[data-test-subj="checkboxSelectRow-${rule.id}"]`).first().prop('checked')
-      ).toBeTruthy();
-    });
-
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    expect(wrapper.find('[data-test-subj="ruleQuickEditButton"]').exists()).toBeTruthy();
-    expect(wrapper.find('[data-test-subj="disableAll"]').first().prop('isDisabled')).toBeTruthy();
-    expect(wrapper.find('[data-test-subj="deleteAll"]').first().prop('isDisabled')).toBeTruthy();
-  });
-
-  it('can bulk snooze', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    // Unselect something to test filtering
-    wrapper.find('[data-test-subj="checkboxSelectRow-2"]').at(1).simulate('change');
-
-    wrapper.find('[data-test-subj="bulkSnooze"]').first().simulate('click');
-
-    expect(wrapper.find('[data-test-subj="snoozePanel"]').exists()).toBeTruthy();
-    wrapper.find('[data-test-subj="linkSnooze1h"]').first().simulate('click');
-
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-
-    expect(bulkSnoozeRules).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ids: [],
-        filter: 'NOT (alert.id: "alert:2")',
-      })
-    );
-  });
-
-  it('can bulk unsnooze', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    // Unselect something to test filtering
-    wrapper.find('[data-test-subj="checkboxSelectRow-2"]').at(1).simulate('change');
-
-    wrapper.find('[data-test-subj="bulkUnsnooze"]').first().simulate('click');
-
-    expect(wrapper.find('[data-test-subj="bulkUnsnoozeConfirmationModal"]').exists()).toBeTruthy();
-    wrapper.find('[data-test-subj="confirmModalConfirmButton"]').first().simulate('click');
-
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-
-    expect(bulkUnsnoozeRules).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ids: [],
-        filter: 'NOT (alert.id: "alert:2")',
-      })
-    );
-  });
-
-  it('can bulk add snooze schedule', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    // Unselect something to test filtering
-    wrapper.find('[data-test-subj="checkboxSelectRow-2"]').at(1).simulate('change');
-
-    wrapper.find('[data-test-subj="bulkSnoozeSchedule"]').first().simulate('click');
-    expect(wrapper.find('[data-test-subj="ruleSnoozeScheduler"]').exists()).toBeTruthy();
-
-    wrapper.find('[data-test-subj="scheduler-saveSchedule"]').first().simulate('click');
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-
-    expect(bulkSnoozeRules).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ids: [],
-        filter: 'NOT (alert.id: "alert:2")',
-      })
-    );
-  });
-
-  it('can bulk remove snooze schedule', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    // Unselect something to test filtering
-    wrapper.find('[data-test-subj="checkboxSelectRow-2"]').at(1).simulate('change');
-
-    wrapper.find('[data-test-subj="bulkRemoveSnoozeSchedule"]').first().simulate('click');
-
-    expect(
-      wrapper.find('[data-test-subj="bulkRemoveScheduleConfirmationModal"]').exists()
-    ).toBeTruthy();
-    wrapper.find('[data-test-subj="confirmModalConfirmButton"]').first().simulate('click');
-
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-
-    expect(bulkUnsnoozeRules).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ids: [],
-        filter: 'NOT (alert.id: "alert:2")',
-        scheduleIds: [],
-      })
-    );
-  });
-
-  it('can bulk update API key', async () => {
-    await setup();
-    wrapper.find('[data-test-subj="checkboxSelectRow-1"]').at(1).simulate('change');
-    wrapper.find('[data-test-subj="selectAllRulesButton"]').at(1).simulate('click');
-    wrapper.find('[data-test-subj="showBulkActionButton"]').first().simulate('click');
-
-    // Unselect something to test filtering
-    wrapper.find('[data-test-subj="checkboxSelectRow-2"]').at(1).simulate('change');
-
-    wrapper.find('[data-test-subj="updateAPIKeys"]').first().simulate('click');
-    expect(wrapper.find('[data-test-subj="updateApiKeyIdsConfirmation"]').exists()).toBeTruthy();
-
-    wrapper.find('[data-test-subj="confirmModalConfirmButton"]').first().simulate('click');
-    await act(async () => {
-      await nextTick();
-      wrapper.update();
-    });
-
-    expect(bulkUpdateAPIKey).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ids: [],
-        filter: 'NOT (alert.id: "alert:2")',
-      })
-    );
   });
 });
